pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Eugen980 <zaza980q@mail.ru>"]
readme = "README.md"
packages = [{ include = "gendiff" }]

[tool.poetry.dependencies]
python = "^3.10"
argparse = "^1.4.0"
flake8 = "^6.1.0"
pytest = "^8.0.2"
pytest-cov = "^4.1.0"
pyyaml = "^6.0.1"

[tool.poetry.group.div.dependencies]

[tool.poetry.scripts]
gendiff = 'gendiff.scripts.gendiff:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
